Extract CustomActionsMixin
